+++ Libraries/MINDful/src/matrices/Matrix3x3f_NoSIMD.cpp
@@ -22,7 +22,7 @@ namespace Mind {
     }
 
     void Matrix3x3f::setColumnValues(
-        const size_t column,
+        const unsigned int column,
         const Point2Df& values
     ) {
         m_data[0][column] = values.get(Vector2f::Axis::X) ;
@@ -30,7 +30,7 @@ namespace Mind {
     }
 
     void Matrix3x3f::setColumnValues(
-        const size_t column,
+        const unsigned int column,
         const Point3Df& values
     ) {
         m_data[0][column] = values.get(Vector3f::Axis::X) ;
@@ -39,7 +39,7 @@ namespace Mind {
     }
 
     void Matrix3x3f::setRowValues(
-        const size_t row,
+        const unsigned int row,
         const Point2Df& values
     ) {
         m_data[row][0] = values.get(Vector2f::Axis::X) ;
@@ -47,7 +47,7 @@ namespace Mind {
     }
 
     void Matrix3x3f::setRowValues(
-        const size_t row,
+        const unsigned int row,
         const Point3Df& values
     ) {
         m_data[row][0] = values.get(Vector3f::Axis::X) ;
@@ -56,7 +56,7 @@ namespace Mind {
     }
 
     Matrix3x3f& Matrix3x3f::operator*=(const Scalar scalar) {
-        unsigned int length = size() ;
+        unsigned int length = static_cast<unsigned int>(size()) ;
         for (unsigned int rowIndex = 0 ; rowIndex < length ; ++rowIndex) {
             for (unsigned int colIndex = 0 ; colIndex < length ; ++colIndex) {
                 m_data[rowIndex][colIndex] *= scalar ;
@@ -68,7 +68,7 @@ namespace Mind {
     Matrix3x3f& Matrix3x3f::operator*=(Matrix3x3f& other) {
         std::array<std::array<float, 4>, 4> result ;
 
-        unsigned int dimension = size() ;
+        unsigned int dimension = static_cast<unsigned int>(size()) ;
         for (unsigned int col = 0 ; col < dimension ; ++col) {
             for (unsigned int row = 0 ; row < dimension ; ++row) {
                 result[row][col] = 0.f ;
@@ -84,7 +84,7 @@ namespace Mind {
     }
 
     Matrix3x3f& Matrix3x3f::operator+=(Matrix3x3f& other) {
-        unsigned int length = size() ;
+        unsigned int length = static_cast<unsigned int>(size()) ;
         for (unsigned int colIndex = 0 ; colIndex < length ; colIndex++) {
             for (unsigned int rowIndex = 0 ; rowIndex < length ; rowIndex++) {
                 m_data[rowIndex][colIndex] += other[rowIndex][colIndex] ;
